pyproject.toml
[tool.poetry]
name = "liveline"
version = "0.1.0"
description = ""
authors = ["RasterMk2 <70229194+RasterMk2@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^2.1.1"
invoke = "^1.7.0"
Flask-Login = "^0.6.0"
python-dotenv = "^0.20.0"
colorama = "^0.4.4"
argon2-cffi = "^21.3.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^22.3.0"
mypy = "^0.942"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
